requirements.txt
tensorflow
opencv-contrib-python 
pandas
numpy
Pillow
tqdm
matplotlib